Move Blacklist, BlacklistEntry, Todolist and TodolistEntry to json package
