+++ src/main/java/com/rettichlp/UnicacityAddon/base/json/BlacklistEntry.java
@@ -1,4 +1,4 @@
-package com.rettichlp.UnicacityAddon.base.json.blacklist;
+package com.rettichlp.UnicacityAddon.base.json;
 
 /**
  * @author RettichLP
