+++ src/main/java/com/rettichlp/UnicacityAddon/base/json/TodolistEntry.java
@@ -1,0 +1,24 @@
+package com.rettichlp.UnicacityAddon.base.json;
+
+public class TodolistEntry {
+
+    private String todo;
+    private boolean done;
+
+    public TodolistEntry(String todo) {
+        this.todo = todo;
+        this.done = false;
+    }
+
+    public String getTodo() {
+        return todo;
+    }
+
+    public boolean isDone() {
+        return done;
+    }
+
+    public void setDone(boolean done) {
+        this.done = done;
+    }
+}
