+++ src/main/java/com/rettichlp/UnicacityAddon/commands/faction/badfaction/ASetBlacklistCommand.java
@@ -2,7 +2,7 @@ package com.rettichlp.UnicacityAddon.commands.faction.badfaction;
 
 import com.rettichlp.UnicacityAddon.base.abstraction.AbstractionLayer;
 import com.rettichlp.UnicacityAddon.base.abstraction.UPlayer;
-import com.rettichlp.UnicacityAddon.base.json.blacklist.BlacklistEntry;
+import com.rettichlp.UnicacityAddon.base.json.BlacklistEntry;
 import com.rettichlp.UnicacityAddon.base.text.ColorCode;
 import com.rettichlp.UnicacityAddon.base.text.Message;
 import com.rettichlp.UnicacityAddon.base.utils.ForgeUtils;
